# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GUG Apis ",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "http://gug-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
...
        "/admin/techniques": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Technique"
                ],
                "summary": "get all techniques",
                "operationId": "91cedf4bca95e89dfc3f92d9a4961ec1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Technique"
                ],
                "summary": "Create a new Technique",
                "operationId": "5d03094392b64169cd5e046611f85598",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTechniqueRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/techniques/{id}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Technique"
                ],
                "summary": "show one Technique",
                "operationId": "8c093820bfd049eeb4668af12e4db143",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Technique"
                ],
                "summary": "update an existing Technique",
                "operationId": "0841b6f95b2dde576b9e51f506df34ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT",
                            "enum": [
                                "PUT"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTechniqueRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin",
                    "Admin - Technique"
                ],
                "summary": "SoftDelete an existing Technique",
                "operationId": "2fc99cac0949e1a654c4449b615977f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
        "/techniques": {
            "get": {
                "tags": [
                    "App",
                    "App - Technique"
                ],
                "summary": "get all techniques",
                "operationId": "fd0ead4a1ef9a3cb35c2358081d10c59",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/techniques/{id}": {
            "get": {
                "tags": [
                    "App",
                    "App - Technique"
                ],
                "summary": "show one Technique",
                "operationId": "a492f241f31378a97fb25bb424596ae2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateTechniqueRequest": {
                "title": "File Upload Request",
                "description": "Request body for uploading a file",
                "properties": {
                    "logo": {
                        "description": "The logo to be uploaded. Accepts image and video formats: jpeg, jpg, png, gif",
                        "type": "string",
                        "format": "binary",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "UpdateTechniqueRequest": {
                "title": "File Upload Request",
                "description": "Request body for uploading a file",
                "properties": {
                    "logo": {
                        "description": "The logo to be uploaded. Accepts image and video formats: jpeg, jpg, png, gif",
                        "type": "string",
                        "format": "binary",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}